+++ selfdrive/car/chrysler/carcontroller.py
@@ -5,14 +5,15 @@ from selfdrive.car.chrysler.values import CAR, CarControllerParams
 from opendbc.can.packer import CANPacker
 from selfdrive.config import Conversions as CV
 
+from selfdrive.controls.lib.drive_helpers import V_CRUISE_MIN, V_CRUISE_MIN_IMPERIAL
 from common.cached_params import CachedParams
 from common.params import Params
 from cereal import car
 import cereal.messaging as messaging
 ButtonType = car.CarState.ButtonEvent.Type
 
-MIN_ACC_SPEED_IMPERIAL_MS = 20 * CV.MPH_TO_MS
-MIN_ACC_SPEED_METRIC_MS = 30 * CV.KPH_TO_MS
+V_CRUISE_MIN_IMPERIAL_MS = V_CRUISE_MIN_IMPERIAL * CV.KPH_TO_MS
+V_CRUISE_MIN_MS = V_CRUISE_MIN * CV.KPH_TO_MS
 AUTO_FOLLOW_LOCK_MS = 3 * CV.MPH_TO_MS
 
 class CarController():
@@ -34,7 +35,7 @@ class CarController():
     self.cachedParams = CachedParams()
     self.disable_auto_resume = self.params.get('jvePilot.settings.autoResume', encoding='utf8') != "1"
     self.autoFollowDistanceLock = None
-    self.minAccSetting = MIN_ACC_SPEED_METRIC_MS if self.params.get_bool("IsMetric") else MIN_ACC_SPEED_IMPERIAL_MS
+    self.minAccSetting = V_CRUISE_MIN_MS if self.params.get_bool("IsMetric") else V_CRUISE_MIN_IMPERIAL_MS
     self.round_to_unit = CV.MS_TO_KPH if self.params.get_bool("IsMetric") else CV.MS_TO_MPH
 
   def update(self, enabled, CS, actuators, pcm_cancel_cmd, hud_alert, gas_resume_speed, jvepilot_state):
@@ -123,15 +124,19 @@ class CarController():
 
   def hybrid_acc_button(self, CS, jvepilot_state):
     # Move the adaptive curse control to the target speed
-    target = jvepilot_state.carControl.vTargetFuture
+    eco_limit = None
     if jvepilot_state.carControl.accEco == 1:  # if eco mode
-      target = min(target, int(CS.out.vEgo + (self.cachedParams.get_float('jvePilot.settings.accEco.speedAheadLevel1', 1000) * CV.MPH_TO_MS)))
+      eco_limit = self.cachedParams.get_float('jvePilot.settings.accEco.speedAheadLevel1', 1000)
     elif jvepilot_state.carControl.accEco == 2:  # if eco mode
-      target = min(target, int(CS.out.vEgo + (self.cachedParams.get_float('jvePilot.settings.accEco.speedAheadLevel2', 1000) * CV.MPH_TO_MS)))
+      eco_limit = self.cachedParams.get_float('jvePilot.settings.accEco.speedAheadLevel2', 1000)
 
-    # keep from fluttering
-    target = round(target * self.round_to_unit) / self.round_to_unit
-    current = round(CS.out.cruiseState.speed * self.round_to_unit) / self.round_to_unit
+    target = jvepilot_state.carControl.vTargetFuture
+    if eco_limit:
+      target = min(target, CS.out.vEgo + (eco_limit * CV.MPH_TO_MS))
+
+    # round to nearest unit
+    target = round(target * self.round_to_unit)
+    current = round(CS.out.cruiseState.speed * self.round_to_unit)
 
     if target < current and current > self.minAccSetting:
       return 'ACC_SPEED_DEC'