+++ RELEASES.md
@@ -1,6 +1,6 @@
 Version jvePilot-0.8.4(v3.3) (2021-XX-XX)
 ========================
-* Better support for metric ACC control
+* Metric ACC control.  Starts at 30kph, ACC +/- changes by 5kph
   
 Version jvePilot-0.8.4(v3.2.2) (2021-05-26)
 ========================